+++ pkg/virt-handler/domain.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"k8s.io/client-go/tools/cache"
 	"kubevirt.io/kubevirt/pkg/logging"
-	"kubevirt.io/kubevirt/pkg/virt-handler/libvirt"
+	"kubevirt.io/kubevirt/pkg/virt-handler/virtwrap"
 )
 
 /*
