+++ pkg/virt-handler/virtwrap/generated_mock_manager.go
